Fixed Medion X10 RF keymap/keytable
After studying both the USB HID Usage Tables and some other remote control drivers I came to the realization that the keytable for the Medion X10 RF is not quite correct. In this commit I changed the keytable so that the remote buttons will match the descriptions of the USB HID Usage Tables as closely as possible. Where in doubt I also checked other remote control drivers to find out how the buttons are assigned to KEY_* constants there in order to get a assignments that match the style of other remotes better.